+++ lib/main.ts
@@ -1,5 +1,6 @@
 import { AdminBar } from '../src/components/AdminBar'
+import { AdminBarBuilder } from '../src/utils/AdminBarBuilder'
 import { AdminBarButton } from '../src/components/AdminBarButton'
 import { AdminBarText } from '../src/components/AdminBarText'
 
-export { AdminBar, AdminBarButton, AdminBarText }
+export { AdminBar, AdminBarBuilder, AdminBarButton, AdminBarText }