+++ vite.config.js
@@ -1,13 +1,17 @@
-import { resolve } from 'path'
+import { dirname, resolve } from 'node:path'
+import { fileURLToPath } from 'node:url'
 import { defineConfig } from 'vite'
+
+const __dirname = dirname(fileURLToPath(import.meta.url))
 
 export default defineConfig({
   build: {
     cssMinify: 'lightningcss',
     lib: {
-      // Could also be a dictionary or array of multiple entry points
-      entry: resolve(__dirname, 'lib/main.ts'),
-      fileName: 'admin-bar',
+      entry: {
+        'admin-bar': resolve(__dirname, 'lib/main.ts'),
+      },
+      fileName: (format, entryName) => `${entryName}.js`,
       formats: ['es'],
     },
     rollupOptions: {